Fix DRY by making a `LspClientBase`
